feat: Implement comprehensive Calculation History feature
- Add calculation history with auto-save functionality
- Create history screen with search, filter, and sort capabilities
- Implement bookmarking for important calculations
- Add statistics view with financial insights
- Support CSV and text export for history data
- Enable sharing calculation summaries
- Store up to 50 recent calculations locally
- Group history by date (Today, Yesterday, This Week, Older)
- Integrate with calculator for seamless load/save operations

+++ lib/core/utils/step_emi_calculation_utils.dart
@@ -67,8 +67,8 @@ class StepEMICalculationUtils {
     }
 
     final averageEMI = steps.isNotEmpty 
-      ? steps.map((s) => s.emiAmount * s.durationMonths).reduce((a, b) => a + b) / (tenureYears * 12)
-      : 0;
+      ? steps.map((s) => s.emiAmount * s.durationMonths).reduce((a, b) => a + b) / (tenureYears * 12).toDouble()
+      : 0.0;
 
     final interestSavedVsRegular = regularTotalInterest - totalInterest;
     final isMoreExpensive = totalInterest > regularTotalInterest;
